+++ src/types.rs
@@ -1,7 +1,10 @@
+mod command;
 mod error;
+mod event;
 mod message;
 
 /// Contains different implementations for `AccessToken` trait.
 pub mod access_token;
 
-pub use {access_token::AccessToken, error::*, message::*};
+pub(crate) use command::*;
+pub use {access_token::AccessToken, error::*, event::*, message::*};