+++ src/types/access_token/basic.rs
@@ -12,10 +12,13 @@ const BASIC: &[u8] = b"Basic ";
 /// # Example
 /// ```rust
 /// # use cometd_client::{types::access_token::Basic, CometdClientBuilder};
+/// # #[tokio::main(flavor = "current_thread")]
+/// # async fn main() {
 /// # let client = CometdClientBuilder::new(&"http://[::1]:1025/".parse().unwrap()).build().unwrap();
 ///
 ///     let access_token = Basic::create("username", Some("password")).unwrap();
 ///     client.update_access_token(access_token);
+/// # }
 /// ```
 #[derive(Debug)]
 pub struct Basic([(&'static str, Box<str>); 1]);