+++ src/types/command.rs
@@ -1,0 +1,6 @@
+#[derive(Debug)]
+pub(crate) enum Command {
+    Handshake,
+    Subscribe(String),
+    Disconnect(String),
+}