+++ src/types/event.rs
@@ -1,0 +1,9 @@
+use crate::types::{CometdError, Data};
+
+#[allow(missing_docs)]
+#[derive(Debug)]
+pub enum Event {
+    Messages(Vec<Data>),
+    Error(CometdError),
+    Disconnected,
+}